requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
numpy==1.24.4
onnx==1.14.0
onnxruntime-gpu==1.15.1
opencv-python==4.8.0.76
av==10.0.0
tqdm==4.65.0
insightface==0.7.3
gfpgan==1.3.8
